move components to widgets
